feat: implement admin authentication middleware and restructure frontend App component
